refactor: split DataReminder into separate file
