+++ plugins/backstage-plugin-coder/src/components/CoderWorkspacesCard/Root.tsx
@@ -4,7 +4,6 @@ import React, {
   useContext,
   useState,
 } from 'react';
-import { makeStyles } from '@material-ui/core';
 import { useId } from '../../hooks/hookPolyfills';
 import { UseQueryResult } from '@tanstack/react-query';
 import {
@@ -16,7 +15,7 @@ import type { Workspace } from '../../typesConstants';
 import { useCoderWorkspacesQuery } from '../../hooks/useCoderWorkspacesQuery';
 import { Card } from '../Card';
 import { CoderAuthWrapper } from '../CoderAuthWrapper';
-import { VisuallyHidden } from '../VisuallyHidden';
+import { DataReminder } from './DataReminder';
 
 type WorkspacesCardContext = Readonly<{
   queryFilter: string;
@@ -28,22 +27,6 @@ type WorkspacesCardContext = Readonly<{
 
 const CardContext = createContext<WorkspacesCardContext | null>(null);
 
-const useStyles = makeStyles(theme => ({
-  button: {
-    color: theme.palette.type,
-    backgroundColor: theme.palette.background.paper,
-    border: 'none',
-    paddingTop: theme.spacing(2),
-    fontSize: theme.typography.body2.fontSize,
-    cursor: 'pointer',
-  },
-
-  snippet: {
-    backgroundColor: theme.palette.grey[100],
-    borderRadius: '0.4em',
-  },
-}));
-
 export type WorkspacesCardProps = Readonly<
   Omit<HTMLAttributes<HTMLDivElement>, 'role' | 'aria-labelledby'> & {
     queryFilter?: string;
@@ -52,42 +35,6 @@ export type WorkspacesCardProps = Readonly<
     readEntityData?: boolean;
   }
 >;
-
-const DataReminder = () => {
-  const [isExpanded, setIsExpanded] = useState(false);
-  const styles = useStyles();
-
-  return (
-    <div>
-      <button
-        type="button"
-        onClick={() => setIsExpanded(!isExpanded)}
-        className={styles.button}
-      >
-        {isExpanded ? '▼' : '►'}{' '}
-        {isExpanded ? 'Hide text' : 'Why am I seeing all workspaces?'}
-      </button>
-
-      {isExpanded && (
-        <p>
-          This component displays all workspaces when the entity has no repo URL
-          to filter by. Consider disabling{' '}
-          <code className={styles.snippet}>readEntityData</code>;{' '}
-          <a
-            href="https://github.com/coder/backstage-plugins/blob/main/plugins/backstage-plugin-coder/docs/components.md#notes-4"
-            rel="noopener noreferrer"
-            target="_blank"
-            style={{ textDecoration: 'underline', color: 'inherit' }}
-          >
-            details in our docs
-            <VisuallyHidden> (link opens in new tab)</VisuallyHidden>
-          </a>
-          .
-        </p>
-      )}
-    </div>
-  );
-};
 
 export const Root = ({
   children,
